HHH-17279 Simplify SqmTranslatorFactory and SqlAstTranslatorFactory
