+++ hibernate-core/src/main/java/org/hibernate/query/sqm/mutation/internal/inline/InlineUpdateHandler.java
@@ -77,8 +77,7 @@ import org.hibernate.sql.ast.tree.update.Assignment;
 import org.hibernate.sql.ast.tree.update.UpdateStatement;
 import org.hibernate.sql.exec.spi.ExecutionContext;
 import org.hibernate.sql.exec.spi.JdbcMutationExecutor;
-import org.hibernate.sql.exec.spi.JdbcOperationQueryInsert;
-import org.hibernate.sql.exec.spi.JdbcOperationQueryUpdate;
+import org.hibernate.sql.exec.spi.JdbcOperationQueryMutation;
 import org.hibernate.sql.exec.spi.JdbcParameterBindings;
 import org.hibernate.sql.results.internal.SqlSelectionImpl;
 
@@ -379,9 +378,9 @@ public class InlineUpdateHandler implements UpdateHandler {
 		final UpdateStatement sqlAst = new UpdateStatement( dmlTableReference, assignments, idListPredicate );
 
 		final JdbcServices jdbcServices = sessionFactory.getJdbcServices();
-		final JdbcOperationQueryUpdate jdbcUpdate = jdbcServices.getJdbcEnvironment()
+		final JdbcOperationQueryMutation jdbcUpdate = jdbcServices.getJdbcEnvironment()
 				.getSqlAstTranslatorFactory()
-				.buildUpdateTranslator( sessionFactory, sqlAst )
+				.buildMutationTranslator( sessionFactory, sqlAst )
 				.translate( jdbcParameterBindings, executionContext.getQueryOptions() );
 
 		final int updateCount = jdbcServices.getJdbcMutationExecutor().execute(
@@ -550,9 +549,9 @@ public class InlineUpdateHandler implements UpdateHandler {
 			insertSqlAst.addTargetColumnReferences( targetColumnReferences.toArray( new ColumnReference[0] ) );
 			insertSqlAst.setSourceSelectStatement( querySpec );
 
-			final JdbcOperationQueryInsert jdbcInsert = jdbcServices.getJdbcEnvironment()
+			final JdbcOperationQueryMutation jdbcInsert = jdbcServices.getJdbcEnvironment()
 					.getSqlAstTranslatorFactory()
-					.buildInsertTranslator( sessionFactory, insertSqlAst )
+					.buildMutationTranslator( sessionFactory, insertSqlAst )
 					.translate( jdbcParameterBindings, executionContext.getQueryOptions() );
 
 			final int insertCount = jdbcServices.getJdbcMutationExecutor().execute(
